redirect working

+++ Implementation/AssistU/app/controllers/DiscussionData.java
@@ -31,8 +31,11 @@ public class DiscussionData extends Controller {
             Project.updateLastAccessed(pid);
             Project p = Project.find.byId(pid);
             return ok(discussion.render("AssistU - Projects", user, p));
-        }else
+        }else {
+            //User did not have a session
+            session().put("callback", routes.DiscussionData.discussion(pid).absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     /** Keeps track of all connected browsers per room **/
@@ -124,30 +127,28 @@ public class DiscussionData extends Controller {
     public static Result deleteMessage() {
         ObjectNode message = (ObjectNode)request().body().asJson();
         User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Comment comment = Comment.find.byId(Long.parseLong(message.get("cid").asText()));
-        Role role = Role.find.where().eq("project", comment.project).eq("user", user).findUnique();
-        Role comment_role = Role.find.where().eq("project", comment.project).eq("user", comment.user).findUnique();
-        List<Comment> comments = Comment.find.where().eq("subject", comment.subject).findList();
-        if(role != null && role.role.equals(Role.OWNER) && comment_role.equals(Role.GUEST)){
-            for(int i = 0; i < comments.size(); i++){
-                comments.get(i).delete();
+        if(user != null) {
+            Comment comment = Comment.find.byId(Long.parseLong(message.get("cid").asText()));
+            Role role = Role.find.where().eq("project", comment.project).eq("user", user).findUnique();
+            Role comment_role = Role.find.where().eq("project", comment.project).eq("user", comment.user).findUnique();
+            List<Comment> comments = Comment.find.where().eq("subject", comment.subject).findList();
+            if (role != null && role.role.equals(Role.OWNER) && comment_role.equals(Role.GUEST)) {
+                for (int i = 0; i < comments.size(); i++) {
+                    comments.get(i).delete();
+                }
+                comment.delete();
+            } else if (comment.user.equals(user)) {
+                for (int i = 0; i < comments.size(); i++) {
+                    comments.get(i).delete();
+                }
+                comment.delete();
             }
-            comment.delete();
-        } else if(comment.user.equals(user)){
-            for(int i = 0; i < comments.size(); i++){
-                comments.get(i).delete();
-            }
-            comment.delete();
         }
         return ok();
     }
 
     public static Result getComments() {
-//        List<Project> ownerProjects = UserData.findActiveOwnerProjects();
-//        List<Project> reviewerProjects = UserData.findActiveReviewerProjects();
-//        List<Project> guestProjects = UserData.findActiveGuestProjects();
         List<Project> projects = UserData.findActiveProjects();
-//        projects.addAll(reviewerProjects);
 
         List<ObjectNode> comments = new ArrayList<ObjectNode>();
         ObjectNode comment;
@@ -172,7 +173,6 @@ public class DiscussionData extends Controller {
                 comments.add(comment);
             }
         }
-//      Logger.debug("Old Comments: " + Json.stringify(toJson(comments)));
         return ok(toJson(comments));
     }
 
@@ -205,7 +205,6 @@ public class DiscussionData extends Controller {
                 comments.add(comment);
             }
         }
-//      Logger.debug("Old SubComments: " + Json.stringify(toJson(comments)));
         return ok(toJson(comments));
     }
