+++ Implementation/AssistU/app/controllers/ProjectData.java
@@ -19,13 +19,16 @@ import static play.libs.Json.toJson;
 public class ProjectData extends Controller {
 
     public static Result project(Long pid) {
+        session().put("callback", routes.ProjectData.project(pid).absoluteURL(request()));
         User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
         if(user != null) {
             Project.updateLastAccessed(pid);
             Project p = Project.find.byId(pid);
             return ok(project.render("AssistU - Projects", user, p));
-        }else
+        } else {
+            session().put("callback", routes.ProjectData.project(pid).absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     static Form<Project> projectForm = Form.form(Project.class);
@@ -34,8 +37,10 @@ public class ProjectData extends Controller {
         User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
         if(user != null)
             return ok(projectNew.render("Create a new Project", projectForm, false, "", "", user));
-        else
+        else {
+            session().put("callback", routes.ProjectData.createProjectPage().absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     public static Result editProjectPage(Long pid) {
@@ -43,8 +48,10 @@ public class ProjectData extends Controller {
         Project p = Project.find.byId(pid);
         if(user != null)
             return ok(projectEdit.render("Edit Project " + p.name, p, projectForm, false, "", "", user));
-        else
+        else {
+            session().put("callback", routes.ProjectData.editProjectPage(pid).absoluteURL(request()));
             return Authentication.login();
+        }
     }
 
     /**
@@ -79,6 +86,7 @@ public class ProjectData extends Controller {
             }
         }
         //User did not have a session
+        session().put("callback", routes.ProjectData.createProject().absoluteURL(request()));
         return Authentication.login();
     }
 
@@ -89,22 +97,28 @@ public class ProjectData extends Controller {
      */
     public static Result editProject(Long pid) {
         User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(ProjectData.findAllOwners(pid).contains(user)) {
-            Project p = Project.find.byId(pid);
-            Form<Project> filledProjectForm = projectForm.bindFromRequest();
-            if (filledProjectForm.hasErrors()) {
-                Logger.debug(filledProjectForm.errors().toString());
-                return badRequest(projectEdit.render("Something went wrong", p, filledProjectForm, true, "danger",
-                        "The input did not fulfill the requirements, please review your information", user));
-            } else if (!(Application.allowedTitleRegex(filledProjectForm.get().folder)
-                    && Application.allowedTitleRegex(filledProjectForm.get().name))) {
-                return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
-                        "The input did not have the allowed format, please review your information", user));
-            } else {
-                Project.edit(pid, filledProjectForm.get().folder, filledProjectForm.get().name, filledProjectForm.get().description);
+        if (user != null) {
+            if (ProjectData.findAllOwners(pid).contains(user)) {
+                Project p = Project.find.byId(pid);
+                Form<Project> filledProjectForm = projectForm.bindFromRequest();
+                if (filledProjectForm.hasErrors()) {
+                    Logger.debug(filledProjectForm.errors().toString());
+                    return badRequest(projectEdit.render("Something went wrong", p, filledProjectForm, true, "danger",
+                            "The input did not fulfill the requirements, please review your information", user));
+                } else if (!(Application.allowedTitleRegex(filledProjectForm.get().folder)
+                        && Application.allowedTitleRegex(filledProjectForm.get().name))) {
+                    return badRequest(projectNew.render("Something went wrong", filledProjectForm, true, "danger",
+                            "The input did not have the allowed format, please review your information", user));
+                } else {
+                    Project.edit(pid, filledProjectForm.get().folder, filledProjectForm.get().name, filledProjectForm.get().description);
+                }
             }
+            return redirect(routes.ProjectData.project(pid));
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.editProject(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.ProjectData.project(pid));
     }
 
     /**
@@ -115,11 +129,17 @@ public class ProjectData extends Controller {
      * @return
      */
     public static Result archiveProject(Long pid) {
-        User u = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(ProjectData.findAllOwners(pid).contains(u) && ProjectData.findAllOwners(pid).size() == 1) {
-            Project.archive(pid);
+        User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            if (ProjectData.findAllOwners(pid).contains(user) && ProjectData.findAllOwners(pid).size() == 1) {
+                Project.archive(pid);
+            }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.archiveProject(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     /*TODO SOHEIL: When a member gets invited, send him a message
@@ -129,34 +149,40 @@ public class ProjectData extends Controller {
      * @param pid: The project to which the user in the form has to be assigned
      * @return
      */
-    public static Result inviteMemberToProjectAs(Long pid){
+    public static Result inviteMemberToProjectAs(Long pid) {
         DynamicForm emailform = Form.form().bindFromRequest();
-        User u = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        //See if the user requesting to invite someone is an Owner
-        if(ProjectData.findAllOwners(pid).contains(u)) {
-            User user = User.find.where().eq("email", emailform.get("email")).findUnique();
-            //See if the user that needs to be invited exists in the system
-            if(user != null) {
-                Project p = Project.find.byId(pid);
-                //There can not be a role relation between the invited user and the project,
-                // as it would be the user is already a member
-                if (Role.find.where().eq("project", p).eq("user", user).findUnique() == null) {
-                    //Pattern match the correct role for invitation
-                    if (emailform.get("role").equals("Owner")) {
-                        Project.inviteOwner(p.id, user.id);
-                        Event.defaultPlanningArticle(user, p);
-                        p.planning = true;
-                        //            p.update();
-                        p.save();
-                    } else if (emailform.get("role").equals("Reviewer")) {
-                        Project.inviteReviewer(p.id, user.id);
-                    } else {
-                        Project.inviteGuest(p.id, user.id);
+        User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if (user != null){
+            //See if the user requesting to invite someone is an Owner
+            if (ProjectData.findAllOwners(pid).contains(user)) {
+                User user_invited = User.find.where().eq("email", emailform.get("email")).findUnique();
+                //See if the user that needs to be invited exists in the system
+                if (user_invited != null) {
+                    Project p = Project.find.byId(pid);
+                    //There can not be a role relation between the invited user and the project,
+                    // as it would be the user is already a member
+                    if (Role.find.where().eq("project", p).eq("user", user_invited).findUnique() == null) {
+                        //Pattern match the correct role for invitation
+                        if (emailform.get("role").equals("Owner")) {
+                            Project.inviteOwner(p.id, user_invited.id);
+                            Event.defaultPlanningArticle(user, p);
+                            p.planning = true;
+                            //            p.update();
+                            p.save();
+                        } else if (emailform.get("role").equals("Reviewer")) {
+                            Project.inviteReviewer(p.id, user_invited.id);
+                        } else {
+                            Project.inviteGuest(p.id, user_invited.id);
+                        }
                     }
                 }
             }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.inviteMemberToProjectAs(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     /*TODO SOHEIL: This function can send a message to all owners that a new user has joined the project
@@ -164,25 +190,37 @@ public class ProjectData extends Controller {
      */
     public static Result hasAccepted(Long pid){
         Project p = Project.find.byId(pid);
-        User u = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Role r = Role.find.where().eq("user", u).eq("project", p).findUnique();
-        //See if there actually exists a (pending) role between the accepting user and project
-        if(r != null){
-            r.accepted=true;
-            r.dateJoined = new Date();
-            r.save();
+        User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if( user != null) {
+            Role r = Role.find.where().eq("user", user).eq("project", p).findUnique();
+            //See if there actually exists a (pending) role between the accepting user and project
+            if (r != null) {
+                r.accepted = true;
+                r.dateJoined = new Date();
+                r.save();
+            }
+            return redirect(routes.ProjectData.project(p.id));
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.hasAccepted(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.ProjectData.project(p.id));
     }
 
     public static Result hasDeclined(Long pid){
         Project p = Project.find.byId(pid);
-        User u = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Role r = Role.find.where().eq("user", u).eq("project", p).findUnique();
-        if(r != null) {
-            r.delete();
+        User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            Role r = Role.find.where().eq("user", user).eq("project", p).findUnique();
+            if (r != null) {
+                r.delete();
+            }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.hasDeclined(pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     /**
@@ -194,17 +232,29 @@ public class ProjectData extends Controller {
      * @return
      */
     public static Result removeMemberFromProject(Long uid, Long pid){
-        User u = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        if(ProjectData.findAllOwners(pid).contains(u)) {
-            Project.removeMemberFrom(pid, uid);
+        User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            if (ProjectData.findAllOwners(pid).contains(user)) {
+                Project.removeMemberFrom(pid, uid);
+            }
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.removeMemberFromProject(uid,pid).absoluteURL(request()));
+            return Authentication.login();
         }
-        return redirect(routes.Application.project());
     }
 
     public static Result leaveProject(Long pid){
-        User u = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
-        Project.removeMemberFrom(pid, u.id);
-        return redirect(routes.Application.project());
+        User user = User.findByAuthUserIdentity(PlayAuthenticate.getUser(session()));
+        if(user != null) {
+            Project.removeMemberFrom(pid, user.id);
+            return redirect(routes.Application.project());
+        } else {
+            //User did not have a session
+            session().put("callback", routes.ProjectData.leaveProject(pid).absoluteURL(request()));
+            return Authentication.login();
+        }
     }
 
     public static Result getProjectIdsAsJson(){
